Add support for hosting files on S3 in production.

+++ app/bin/entrypoint.sh
@@ -7,12 +7,11 @@
 #  - Django admin user is not created as part of this.
 #  - No special behaviour to be added for the initial run. That is outside this scripts responsibilities.
 #  - Set `DJANGO_DEBUG` to trigger installation of Python dev dependencies.
+#    - Installation of dev dependencies only occurs if Pipfile.lock is not present.
 #  - Set `DJANGO_MANAGEPY_MIGRATE` to trigger migrations.
 #  - Set `DJANGO_MANAGEPY_COLLECTSTATIC` to trigger the collection of static files.
 #
 #--------------------------------------------
-
-# TODO: Move this into the wsgi.py file so application code isn't maintained in multiple places.
 
 set -e
 
@@ -50,8 +49,10 @@ fi
 #-------------------------------------------------------------------------------
 
 if [ -v DJANGO_DEBUG ]; then
-    echo '[info]: Installing development packages as DJANGO_DEBUG is set.'
-    pipenv install --dev --deploy --system
+    if [ ! -f Pipfile.lock ]; then
+        echo '[info]: Installing development packages as DJANGO_DEBUG is set.'
+        pipenv install --dev --deploy --system
+    fi
 fi
 
 if [ -v DJANGO_MANAGEPY_MIGRATE ]; then
